configure database data

+++ database/migrations/2020_12_09_085832_create_order_details_table.php
@@ -16,8 +16,8 @@ class CreateOrderDetailsTable extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('order_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('order_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->integer('qty');
             $table->double('amount');
             $table->double('total');
